\name{PopMCarma-package}
\alias{PopMCarma-package}
\alias{PopMCarma}
\docType{package}
\title{
\packageTitle{PopMCarma}
}
\description{
\packageDescription{PopMCarma}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{PopMCarma}
\packageIndices{PopMCarma}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{PopMCarma}

Maintainer: \packageMaintainer{PopMCarma}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
